massive server improvement

+++ jtogether_server/model/user-model.js
@@ -1,20 +1,36 @@
 const userApi = require('../apis/user-api');
-const jwt = require('jsonwebtoken');
-
 
 const USER_NOT_FOUND = 'Utente non trovato, username: '
 const CREDENTIAL_ALREADY_PRESENT = 'Username o email già presenti'
+const USER_MUST_DELETE_ACTIVITIES = 'Cancellare prima tutte le attività'
 
 module.exports = {
     deleteUser,
-    updateUser
+    updateUser,
+    deleteParticipation,
+    createParticipation,
+    deleteActivity,
+    createActivity,
+    getUserFromUsername,
+    getUserFromEmail,
+    createUser
 }
 
 async function deleteUser({username}){
-    const user = await userApi.deleteUser(username)
+    const user = await userApi.getUserFromUsername(username)
+
     if(!user){
         throw USER_NOT_FOUND + username
     }
+    if(user.activities_created.length > 0){
+        throw USER_MUST_DELETE_ACTIVITIES
+    }
+    await userApi.deleteUser(username)
+    return user.activities_participated
+}
+
+async function createUser(userParams){
+    return userApi.createUser(userParams)
 }
 
 async function updateUser(userParams){
@@ -24,6 +40,29 @@ async function updateUser(userParams){
     if(userFromEmail || userFromUsername){
         throw CREDENTIAL_ALREADY_PRESENT
     }
-
     return (await userApi.updateUser(userParams)).toJSON()
+}
+
+async function deleteParticipation({username},{activity_id}){
+    return userApi.deleteParticipation(username,activity_id)
+}
+
+async function createActivity({username},{activity_id}){
+    return userApi.createActivity(username,activity_id)
+}
+
+async function deleteActivity({username},{activity_id}){
+    return userApi.deleteActivity(username,activity_id)
+}
+
+async function createParticipation({username},{activity_id}){
+    return userApi.createParticipation(username,activity_id)
+}
+
+async function getUserFromUsername({username}){
+    return userApi.getUserFromUsername(username)
+}
+
+async function getUserFromEmail({email}){
+    return userApi.getUserFromEmail(email)
 }
